refactor(ai-gateway): group partner providers
