+++ apps/web/src/lib/ai-gateway/custom-pricing.ts
@@ -2,8 +2,7 @@ import { captureMessage } from '@sentry/nextjs';
 import type { OpenRouterModel } from '@/lib/organizations/organization-types';
 import type { JustTheCostsUsageStats } from '@/lib/ai-gateway/processUsage.types';
 import { QWEN37_MAX_MODEL_ID, QWEN37_PLUS_MODEL_ID } from '@/lib/ai-gateway/providers/qwen';
-import { PERPLEXITY_KIMI_PUBLIC_ID } from '@/lib/ai-gateway/providers/moonshotai';
-import { FRIENDLI_GLM_PUBLIC_ID } from '@/lib/ai-gateway/providers/zai';
+import { partnerPricingByModelId } from '@/lib/ai-gateway/providers/partner/pricing';
 import {
   calculateCost_mUsd,
   type Pricing,
@@ -61,34 +60,7 @@ export const customPricingByModelId: Record<string, CustomPricing> = {
       },
     ],
   },
-  [PERPLEXITY_KIMI_PUBLIC_ID]: {
-    fallbackOnly: true,
-    pricing: [
-      {
-        start_context_length: 0,
-        pricing: {
-          prompt_per_million: 3,
-          completion_per_million: 15,
-          input_cache_read_per_million: 0.3,
-          input_cache_write_per_million: null,
-        },
-      },
-    ],
-  },
-  [FRIENDLI_GLM_PUBLIC_ID]: {
-    fallbackOnly: true,
-    pricing: [
-      {
-        start_context_length: 0,
-        pricing: {
-          prompt_per_million: 1.4,
-          completion_per_million: 4.4,
-          input_cache_read_per_million: 0.26,
-          input_cache_write_per_million: null,
-        },
-      },
-    ],
-  },
+  ...partnerPricingByModelId,
 };
 
 export function getCustomPricing(modelId: string): CustomPricing | undefined {
