+++ apps/web/src/lib/ai-gateway/providers/partner/constants.ts
@@ -1,0 +1,3 @@
+// Partner routing stays pinned when the current public models change.
+export const FRIENDLI_GLM_PUBLIC_ID = 'z-ai/glm-5.2';
+export const PERPLEXITY_KIMI_PUBLIC_ID = 'moonshotai/kimi-k3';
